Updated observation operator setup for local distributions

+++ src/oops/generic/hdiag_nicas/type_randgen_api.cpp
@@ -10,7 +10,7 @@
 using namespace std;
 
 // Constructor
-RANDGEN* create_randgen(int default_seed){
+RANDGEN* create_randgen(unsigned long int default_seed){
     return new randGen(default_seed);
 }
 
@@ -20,7 +20,7 @@ void delete_randgen(RANDGEN* randgen){
 }
 
 // Reseed generator
-void reseed_randgen(RANDGEN* randgen, int seed) {
+void reseed_randgen(RANDGEN* randgen, unsigned long int seed) {
     randgen->reseed_randgen(seed);
 }
